+++ cli/src/pcluster/templates/cluster_stack.py
@@ -830,13 +830,6 @@ class ClusterCdkStack(Stack):
                 ),
                 tag_specifications=[
                     ec2.CfnLaunchTemplate.TagSpecificationProperty(
-                        resource_type="instance",
-                        tags=get_default_instance_tags(
-                            self._stack_name, self.config, head_node, "HeadNode", self.shared_storage_infos
-                        )
-                        + get_custom_tags(self.config),
-                    ),
-                    ec2.CfnLaunchTemplate.TagSpecificationProperty(
                         resource_type="volume",
                         tags=get_default_volume_tags(self._stack_name, "HeadNode") + get_custom_tags(self.config),
                     ),
@@ -1110,6 +1103,10 @@ class ClusterCdkStack(Stack):
                 launch_template_id=head_node_launch_template.ref,
                 version=head_node_launch_template.attr_latest_version_number,
             ),
+            tags=get_default_instance_tags(
+                self._stack_name, self.config, head_node, "HeadNode", self.shared_storage_infos
+            )
+            + get_custom_tags(self.config),
         )
         if not self._condition_is_batch():
             head_node_instance.node.add_dependency(self.compute_fleet_resources)